v0.6.1
Adds a button for copying the ticket info with a link that can be pasted into another application

Additonaly modificaiton to how most of the buttons are added to the pages was performed. Its now using a page change event so that the various buttons are removed when they wont function and are added to the page even if a full page refresh isnt performed.